feat: Implement user authentication and database-backed settings
This commit introduces a complete user authentication and management
system with a PostgreSQL database backend.

Key Features:
- User sign-up and sign-in with email/password and Google Sign-In.
- Two user roles: 'admin' and 'user'.
- Admin users can view and manage all users from a new admin dashboard.
- User settings, previously stored in a local `.midiatorsetup` file, are
  now saved to a `settings` table in the database, linked to each user.
- The application is refactored to use React Router with protected
  routes based on authentication status and user role.

Technical Changes:
- Adds backend API endpoints for auth, user management, and settings.
- Introduces a database schema for `users` and `settings` tables.
- Adds required dependencies for backend (`pg`, `bcrypt`, `jsonwebtoken`)
  and frontend (`react-router-dom`, `@react-oauth/google`).
- Refactors the main application into a routed application with separate
  pages for login, signup, home, and admin dashboard.
- Implements a `UserAuthContext` for global state management of the
  authenticated user.
