+++ api/auth/logout.js
@@ -1,0 +1,21 @@
+import { serialize } from 'cookie';
+
+export default async function handler(req, res) {
+  if (req.method !== 'POST') {
+    res.setHeader('Allow', ['POST']);
+    return res.status(405).end(`Method ${req.method} Not Allowed`);
+  }
+
+  // To log out, we clear the cookie by setting its maxAge to -1.
+  // The browser will then delete the cookie.
+  const cookie = serialize('auth_token', '', {
+    httpOnly: true,
+    secure: process.env.NODE_ENV !== 'development',
+    sameSite: 'strict',
+    maxAge: -1, // Expire the cookie immediately
+    path: '/',
+  });
+
+  res.setHeader('Set-Cookie', cookie);
+  res.status(200).json({ message: 'Logged out successfully.' });
+}
