+++ package.json
@@ -21,29 +21,36 @@
     "@mui/material": "^5.14.20",
     "@mui/x-data-grid": "^6.18.2",
     "@mui/x-date-pickers": "^8.10.0",
+    "@react-oauth/google": "^0.12.2",
     "@uppy/core": "^4.4.7",
     "@uppy/react": "^4.4.0",
     "@uppy/screen-capture": "^4.3.1",
+    "bcryptjs": "^3.0.2",
     "chroma-js": "^3.1.2",
     "clsx": "^2.1.1",
     "colorthief": "2.6.0",
+    "cookie": "^1.0.2",
     "crypto-js": "^4.2.0",
     "date-fns": "^4.1.0",
     "file-saver": "^2.0.5",
     "gapi-script": "^1.2.0",
+    "google-auth-library": "^10.2.1",
     "html-react-parser": "^5.2.6",
     "html2canvas": "^1.4.1",
     "ioredis": "^5.7.0",
     "jose": "^6.0.12",
+    "jsonwebtoken": "^9.0.2",
     "jszip": "^3.10.1",
     "lodash.isequal": "^4.5.0",
     "node-fetch": "^3.3.2",
     "pako": "^2.1.0",
     "papaparse": "^5.4.1",
+    "pg": "^8.16.3",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-draggable": "^4.5.0",
+    "react-router-dom": "^7.8.1",
     "sonner": "^1.1.2",
     "tailwind-merge": "^3.3.1",
     "vaul": "^1.1.2"
