+++ setup.sql
@@ -1,0 +1,49 @@
+-- Function to automatically update the updated_at timestamp
+CREATE OR REPLACE FUNCTION trigger_set_timestamp()
+RETURNS TRIGGER AS $$
+BEGIN
+  NEW.updated_at = NOW();
+  RETURN NEW;
+END;
+$$ LANGUAGE plpgsql;
+
+-- Users Table
+CREATE TABLE IF NOT EXISTS users (
+    id SERIAL PRIMARY KEY,
+    uuid UUID DEFAULT gen_random_uuid() NOT NULL UNIQUE,
+    name VARCHAR(255),
+    email VARCHAR(255) UNIQUE NOT NULL,
+    password_hash VARCHAR(255),
+    role VARCHAR(50) DEFAULT 'user' NOT NULL,
+    google_id VARCHAR(255) UNIQUE,
+    created_at TIMESTAMPTZ DEFAULT NOW(),
+    updated_at TIMESTAMPTZ DEFAULT NOW()
+);
+
+-- Trigger for users table
+DROP TRIGGER IF EXISTS set_timestamp ON users;
+CREATE TRIGGER set_timestamp
+BEFORE UPDATE ON users
+FOR EACH ROW
+EXECUTE PROCEDURE trigger_set_timestamp();
+
+-- Settings Table
+CREATE TABLE IF NOT EXISTS settings (
+    id SERIAL PRIMARY KEY,
+    user_id INTEGER NOT NULL UNIQUE REFERENCES users(id) ON DELETE CASCADE,
+    settings_data JSONB,
+    created_at TIMESTAMPTZ DEFAULT NOW(),
+    updated_at TIMESTAMPTZ DEFAULT NOW()
+);
+
+-- Trigger for settings table
+DROP TRIGGER IF EXISTS set_timestamp ON settings;
+CREATE TRIGGER set_timestamp
+BEFORE UPDATE ON settings
+FOR EACH ROW
+EXECUTE PROCEDURE trigger_set_timestamp();
+
+-- Add a comment indicating successful execution
+-- This is a simple way to verify the script ran without errors
+COMMENT ON TABLE users IS 'Users table created successfully';
+COMMENT ON TABLE settings IS 'Settings table created successfully';
